+++ test/dummy/app/notifiers/application_notifier.rb
@@ -1,2 +1,6 @@
 class ApplicationNotifier < Noticed::Event
+  notification_methods do
+    def inherited_method
+    end
+  end
 end